refactor: deprecate baseEntitySheet and implement documentSheet
