+++ index.d.ts
@@ -95,6 +95,7 @@ import './src/foundry/applications/formApplications/baseEntitySheet';
 import './src/foundry/applications/formApplications/combatantConfig';
 import './src/foundry/applications/formApplications/combatTrackerConfig';
 import './src/foundry/applications/formApplications/drawingConfig';
+import './src/foundry/applications/formApplications/documentSheet';
 import './src/foundry/applications/formApplications/entitySheetConfig';
 import './src/foundry/applications/formApplications/folderConfig';
 import './src/foundry/applications/formApplications/gridConfig';
